migration_engine/requirements.txt
anthropic>=0.40.0
typer>=0.12.0
rich>=13.0.0
python-dateutil>=2.9.0
typing-extensions>=4.0.0
